Initial stab at porting `asm/stack.ts` to Rust
This commit is an initial stab at adding some Rust to glimmer, notably to
implement some bits-and-pieces of the `low-level` package for now. Specifically
this commit takes the `Stack` type in the `low-level` module and ports it to
Rust and then integrates it all back with WebAssembly.

Some notable points of this:

* The error handling isn't great right now and the interface may not be quite
  right just yet.
* Right now the raw wasm instructions are used to allocate memory (the Rust
  standard library's allocator is avoided) and internally a freelist is
  maintained of free'd pages.
* Currently memory is never deallocated in Rust which means that pages are
  forever in use and usage will "grow without bound". Eventually a manual call
  to "free this" will need to be inserted for the `Stack` type.